+++ qml/pages/CalcButton.qml
@@ -1,5 +1,6 @@
 import QtQuick 2.2
 import Sailfish.Silica 1.0
+import QtFeedback 5.0
 
 Button
 {
@@ -38,7 +39,7 @@ Button
 	function backspace()
 	{
 		var pos = textfield.cursorPosition;
-		if ( textfield.text == "" || pos == 0 )
+		if ( textfield.text == "" || pos === 0 )
 			return;
 		if ( textfield.selectionStart - textfield.selectionEnd != 0 )
 		{
@@ -77,17 +78,27 @@ Button
 		return true
 	}
 
+	ThemeEffect
+	{
+		id: pressEffect
+		effect: "Press"
+	}
+
 	onPressed: { screen.interactive = false; keyboard.interactive = false }
 	onReleased: { screen.interactive = true; keyboard.interactive = true }
 	onExited: { screen.interactive = true; keyboard.interactive = true }
 	onCanceled: { screen.interactive = true; keyboard.interactive = true }
 	onClicked:
 	{
+		if ( hapticfeedback && pressEffect.supported )
+			pressEffect.play()
 		if ( !checkmacro(value) )
 			insertValue(value)
 	}
 	onPressAndHold:
 	{
+		if ( hapticfeedback && pressEffect.supported )
+			pressEffect.play()
 		if ( !checkmacro(second) )
 			insertValue(second)
 	}